+++ app/server.js
@@ -13,6 +13,7 @@ require('express-async-errors');
 // Routes
 const user = require('./routes/user');
 const portfolio = require('./routes/portfolio');
+const game = require('./routes/game');
 
 
 let iex = require('./utils/iex');
@@ -114,6 +115,7 @@ app.use(express.static('public'));
 
 app.use("/user", user);
 app.use("/portfolio", portfolio);
+app.use("/game", game);
 
 
 
@@ -248,40 +250,6 @@ app.get("/cash", (req, res) => {
         return res.json(result);
     });
 })
-
-
-app.get(`/game/create`, (req, res) => {
-    res.render('create_games');
-})
-
-app.post("/game/create", asyncHandler(async (req, res) => {
-    let userId = Number(req.user.user_id);
-    let gameName = req.body.name;
-    let cash = Number(req.body.cash);
-    let userList = req.body.users;
-    // console.log("User ID:",userId);
-    // console.log("Game name:",gameName);
-    // console.log("Cash:",cash);
-    // console.log("MyUser:",req.user.username);
-    // console.log("Users:",req.body.users);
-
-    //First create game from input, then use that game-ID to create a portfolio and link the game-ID with the user-ID of creator
-    //Then loop through the list of all users that should be in the game and do the same
-    await db.createGame(pool, userId, gameName, cash);
-    gameIdResponse = await db.getGameId(pool, gameName);
-    gameId = gameIdResponse[0].game_id;
-    await db.createPortfolio(pool, userId, gameName + " " + req.user.username, cash, gameId);
-    await db.linkGame(pool, userId, gameId);
-
-    for (i in userList) {
-        userIdResponse = await db.getUserId(pool, userList[i]);
-        specificUserId = userIdResponse[0].user_id;
-        await db.createPortfolio(pool, specificUserId, gameName + " " + userList[i], cash, gameId);
-        await db.linkGame(pool, specificUserId, gameId);
-        console.log("created portfolio for", userList[i]);
-    }
-    return res.redirect('/game/create');
-}));
 
 app.listen(port, hostname, () => {
     console.log(`http://${hostname}:${port}`);
